monitoring service - not tested yet!

+++ src/main/java/tuwien/aic12/server/dao/CustomerDao.java
@@ -45,7 +45,7 @@ public class CustomerDao implements Dao<Customer> {
     }
 
     @Override
-    public Customer read(int id) {
+    public Customer read(long id) {
         return em.find(Customer.class, id);
     }
